+++ src/common/helpers/json-content-required.helper.ts
